Support React 16 as a peer and dev dependency
This change moves the dev dependency from React 15.6, to React 16,
and adds React 16 as an accepted peer dependency. Comparing the differences
between 15.6 and 16 APIs, there are no breaking changes affecting React Tracking.

On the dev side Enzyme has been bumped to 3.x and setup to use its React 16 adapter.
There is also an added polyfill for requestAnimiationFrame support, which 16 relies on,
and which isn't available in the test environment.

This change does not address https://github.com/NYTimes/react-tracking/issues/67 at all,
but acts as a starting point to it.